I worked on creating the image-store-context. Updated actions
The image store is behind a react context that will store
the current fetched images, this will be useful to
optimize the api calls. I don't need router.refresh anymore
I just need to return the imageId from a delete operation
or a image data from a insert operation and modify the
image store on the client.

I implemented the base for the infinite schorl to work
using the image store, for now it works with a button but
the next stem is to auto detect when user schorl to the bottom
and fetch more images based on the last image createdAt field
from the images from image store.

I updated the actions, the delete will return the imageId now
and the confirmImage will return a ImageData that will be
inserted at the start of the image store.